Refactor Height + Weight Questions (#362)
* refactor height question

* refactor weight question

* remove unused comment lines

* fix imports
